diva-embedded/training/wakeword/requirements.txt
# === Core ===
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
pyyaml>=6.0
tqdm>=4.64.0
requests>=2.28.0

# === TTS ===
piper-tts>=1.2.0

# === Audio augmentation ===
audiomentations>=0.33.0

# === PyTorch (CUDA 12.1 pour RTX 4090) ===
# Installer séparément :
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchaudio>=2.0.0
torchmetrics>=1.2.0

# === ONNX (versions critiques pour la conversion) ===
onnx==1.17.0
onnxruntime-gpu==1.18.1

# === Conversion TFLite ===
tensorflow==2.19.0
onnx2tf

# === Entraînement openWakeWord ===
mutagen==1.47.0
torchinfo==1.8.0
speechbrain==0.5.14
acoustics==0.2.6
pronouncing==0.2.0
datasets>=2.14.6
deep-phonemizer==0.0.19

# === Enregistrement micro ===
pyaudio>=0.2.13

# === openWakeWord ===
openwakeword>=0.6.0
